Been lazy.
Working on figuring out how to handle all the caches in a optimal and visually appealing way.

Changed "Thread" isEnabled to isBusy because isEnabled was a bad choice it's to check if the thread is currently running not if it's enabled like a control similar to  bw.IsBusy property

Fixed a random crash when json wasn't parsing right and I left a Throw in.